% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featimportance.R
\name{important_mercedes_features}
\alias{important_mercedes_features}
\title{Return features important to the determination of predicted mercedes pricing.}
\usage{
important_mercedes_features(data, model, ascending = FALSE)
}
\arguments{
\item{data}{input data with columns (i.e. "features"). Should be the same data used to train the prediction model.}

\item{model}{model used to create the predicted mercedes price.}

\item{ascending}{when ascending=FALSE, feature importances are returned from highest to lowest importance.}
}
\value{
a visualization chart/table of sorted feature importances.
}
\description{
Return features important to the determination of predicted mercedes pricing.
}
\examples{
feature_importance_plot <- important_mercedes_features(X_train, model_XGboost)
}
