% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary.R
\name{listing_summary}
\alias{listing_summary}
\title{Summarize mileage, price, and rarity information}
\usage{
listing_summary(
  data,
  model = "all",
  model_col = "model",
  odometer_col = "odometer_mi",
  price_col = "price_USD"
)
}
\arguments{
\item{data}{data.frame The input Mercedes-Benz dataframe}

\item{model}{character A character vector that lists all Mercedes-Benz model types to be included in the final summary table.}

\item{model_col}{character The name of the column in the input dataframe that represents the Mercedes-Benz model}

\item{odometer_col}{character The name of the column in the input dataframe that represents the Mercedes-Benz mileage (odometer reading in miles)}

\item{price_col}{character The name of the column in the input dataframe that represents the Mercedes-Benz sale price (price in USD)}
}
\value{
data.frame Summary dataframe of Mercedes-Benz models, including average and median mileage (mi) and price (USD), and information about rarity. A model will be considered common, uncommon, or rare based on the percentage of market listings.
}
\description{
Returns a dataframe summary of all existing Mercedes-Benz car listings
Summary mileage, price, and rarity information is shown.
}
\examples{
toy_data <- data.frame(model = c("gl-class", "gl-class", "gl-class", "amg", "amg", "amg"),
odometer_mi = c(100000, 200000, 400000, 100000, 200000, 400000),
price_USD = c(10000, 20000, 40000, 10000, 20000, 40000))
listing_summary(data = toy_data, model="all")
listing_summary(data = toy_data, model="gl-class")
}
