% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{data}
\name{mercedes_data}
\alias{mercedes_data}
\title{Mercedes-Benz Used-car Dataset}
\format{
\subsection{\code{mercedes_data}}{

A data frame with 5,553 rows and 17 columns:
\describe{
\item{price_USD}{Price of car in US dollar}
\item{condition}{Condition of the car}
\item{paint_color}{paint color}
\item{model}{model of the car}
\item{odometer_mi}{odometer (in miles) of the car}
\item{year}{year of production}
\item{num_cylinders}{number of cylinders of the car}
\item{fuel}{fuel needed of the car}
\item{transmission}{transmission of the car}
\item{drive}{drive of the car}
\item{size}{size of the car}
\item{type}{type of the car}
\item{state}{location (state) of the car}
\item{VIN}{VIN number of the car}
\item{title_status}{title status of the car}
\item{description}{description of the car}

}
}
}
\source{
\url{https://github.com/AustinReese/UsedVehicleSearch}
}
\usage{
mercedes_data
}
\description{
A subset of data from the Used-car data
}
\keyword{datasets}
