% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mercedes_price_model}
\alias{mercedes_price_model}
\title{Mercedes-Benz Used-car Price Prediction Model}
\format{
\subsection{\code{mercedes_price_model}}{

A model object.
}
}
\usage{
mercedes_price_model
}
\description{
A script for loading the pre-trained model for price prediction
}
\keyword{datasets}
