% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_mercedes_price}
\alias{plot_mercedes_price}
\title{Plot Mercedes Price}
\usage{
plot_mercedes_price(model, price, market_df)
}
\arguments{
\item{model}{character The model of the vehicle.}

\item{price}{numeric  The price of the vehicle.}

\item{market_df}{data.frame Data containing information on used Mercedes-Benz in the market.}
}
\value{
ggplot Density plot of prices for the specified Mercedes-Benz model.
}
\description{
Plot a density plot of a specific Mercedes-Benz model to see where the current vehicle's price
falls within the distribution of prices for that model in the market.
}
\examples{
plot_mercedes_price(model='S-Class', price=80000, market_df=market_df))
plot_mercedes_price(model='C-Class', price=10000, market_df=used_car_df))
}
