% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{listing_search}
\alias{listing_search}
\title{Return the top listings that are within the budget specified by the user.
The results are filtered by an optional input, model, and sorted by ascending price and another the specified feature in the sort_feature parameter.}
\usage{
listing_search(
  data,
  budget = c(0, Inf),
  model = "any",
  sort_feature = "odometer",
  ascending = TRUE
)
}
\arguments{
\item{data}{data.frame The input dataframe containing listing and car attributes}

\item{budget}{number or numeric vector of length 2 The maximum budget, or the range of budget when passing a list.}

\item{model}{character The model of the car that the user is interested in. The default is to include listings of any model.}

\item{sort_feature}{character The numeric variable that the user is interested in using to sort the result. The default value is to sort by odometer value.}

\item{ascending}{bool Boolean value that indicate whether the sort is ascending. The default value is True.}
}
\value{
data.frame A data frame of the sorted listings that matches user's expected budget range.
}
\description{
Return the top listings that are within the budget specified by the user.
The results are filtered by an optional input, model, and sorted by ascending price and another the specified feature in the sort_feature parameter.
}
\examples{
listing_search(data, budget=c(0, 30000), model = "any", sort_feature = "odometer", ascending = TRUE)
listing_search(data, budget=50000, model = "gls", sort_feature = "year", ascending = FALSE)
}
